% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dagsterpipes-package.R
\docType{package}
\name{dagsterpipes-package}
\alias{dagsterpipes}
\alias{dagsterpipes-package}
\title{dagsterpipes: Dagster Pipes Protocol for R}
\description{
Implements the Dagster Pipes protocol, enabling R scripts to communicate with the Dagster orchestrator. R scripts can receive execution context and report asset materializations, check results, and log messages back to Dagster.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/joekirincic/dagsterpipes}
  \item \url{https://joekirincic.github.io/dagsterpipes/}
  \item Report bugs at \url{https://github.com/joekirincic/dagsterpipes/issues}
}

}
\author{
\strong{Maintainer}: Joseph Kirincic \email{jdkirincic@gmail.com}

}
\keyword{internal}
