% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{json_serialize}
\alias{json_serialize}
\title{Serialize an R object to Dagster-compatible JSON}
\usage{
json_serialize(obj)
}
\arguments{
\item{obj}{An R object to serialize.}
}
\value{
A JSON string.
}
\description{
Serialize an R object to Dagster-compatible JSON
}
\keyword{internal}
