% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/message-writer.R
\name{PipesFileMessageWriterChannel}
\alias{PipesFileMessageWriterChannel}
\title{Pipes File Message Writer Channel}
\description{
Pipes File Message Writer Channel

Pipes File Message Writer Channel
}
\details{
Writes newline-delimited JSON messages to a temporary file for the Dagster
Pipes protocol.
}
\keyword{internal}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PipesFileMessageWriterChannel-new}{\code{PipesFileMessageWriterChannel$new()}}
\item \href{#method-PipesFileMessageWriterChannel-write_message}{\code{PipesFileMessageWriterChannel$write_message()}}
\item \href{#method-PipesFileMessageWriterChannel-clone}{\code{PipesFileMessageWriterChannel$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PipesFileMessageWriterChannel-new"></a>}}
\if{latex}{\out{\hypertarget{method-PipesFileMessageWriterChannel-new}{}}}
\subsection{Method \code{new()}}{
Create a new message writer channel.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PipesFileMessageWriterChannel$new(path)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{Path to the messages file.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PipesFileMessageWriterChannel-write_message"></a>}}
\if{latex}{\out{\hypertarget{method-PipesFileMessageWriterChannel-write_message}{}}}
\subsection{Method \code{write_message()}}{
Write a message to the file.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PipesFileMessageWriterChannel$write_message(message)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{message}}{A list representing a Dagster Pipes message.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PipesFileMessageWriterChannel-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PipesFileMessageWriterChannel-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PipesFileMessageWriterChannel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
